% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_fits.R
\name{apply_fits}
\alias{apply_fits}
\title{Run a given function on a simpr_mod object and
tidy the output.}
\usage{
apply_fits(obj, .f, ...)
}
\arguments{
\item{obj}{tibble with repetition number,
metaparameters, simulated data, and fitted
models, from \code{\link{fit}}}

\item{.f}{A function or \code{purrr}-style formula function used for computing on the
fit object}

\item{\dots}{Additional arguments to \code{.f}.}
}
\description{
This function applies a given function to all
fit objects and returns the result in a tidy
tibble.  Any function or \code{purrr}-style
lambda function can be used.
}
\seealso{
\code{\link{tidy_fits}}, \code{\link{glance_fits}}
}
