% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sims.R
\name{as_sims}
\alias{as_sims}
\alias{as_tibble.simpr_sims}
\title{Work directly with simulation results with
dplyr and tidyr}
\usage{
as_sims(obj)

as_tibble.simpr_sims(obj)
}
\arguments{
\item{obj}{A \code{simpr_tibble} or \code{simpr_spec} object.}
}
\value{
A \code{simpr_sims} object for use with dplyr and tidyr verbs.
}
\description{
Allows applying data transformations to every
simulation result with syntax as if dealing
with a single simulation result using dplyr and
tidyr verbs
}
\details{
After producing simulation results (a
\code{simpr_tibble} object), it is sometimes
needed to do some data transformation to
prepare for analysis.  This can always be
specified in blueprint through custom
functions, but \code{as_sims} allows you to
also easily specify this in your pipeline.
After running as_sims, you can use the dplyr
and tidyr verbs you would use on a single
simulation result and it will be applied to all
results.

If, after running \code{as_sims}, you wish to
return to the default behavoir to access
\code{simpr_tibble} results as a tibble with a
list_column for simulation results again, run
\code{as_tibble()}.
}
