% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{as_tibble.simpr_sims}
\alias{as_tibble.simpr_sims}
\alias{simpr_tibble}
\title{Convert a simpr_sims object back to a
simpr_tibble or bare tibble}
\usage{
\method{as_tibble}{simpr_sims}(x, ..., bare_tibble = FALSE, .rows, .name_repair, rownames)
}
\arguments{
\item{x}{A \code{simpr_sims} or
\code{simpr_tibble} object.}

\item{...}{Ignored.}

\item{bare_tibble}{Convert to a bare tibble}

\item{.rows}{Ignored.}

\item{.name_repair}{Ignored.}

\item{rownames}{Ignored.}
}
\value{
A tibble with the metaparameters and simulation results
}
\description{
Allows access to simulation results as a
tibble, with simulations available as a
list-column.
}
\details{
This function is the inverse of
\code{\link{per_sim}}.  This enables
tidyverse verbs to act on the full table as
opposed to the default behavior of acting
elementwise on the simulation results. If
\code{bare_tibble = TRUE}, the output is a bare
tibble object that will not work with simpr
methods such as \code{\link{fit}}.
}
