% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpr_fxns.R
\name{calc_glance}
\alias{calc_glance}
\title{Glance at the simulated model results output (broom analogue)}
\usage{
calc_glance(simpr_mod)
}
\arguments{
\item{simpr_mod}{simulated model results (output of simpr::fit)}
}
\value{
tidied model summary of fitted simulated data
}
\description{
Turn fitted model of simulated data (output of simpr::fit) into a tidy tibble of overall model statistics
}
\examples{
calc_glance(simpr_mod)

}
