% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpr_fxns.R
\name{calc_tidy}
\alias{calc_tidy}
\title{Tidy the simulated model results output into tibble of components (broom analogue)}
\usage{
calc_tidy(simpr_mod)
}
\arguments{
\item{simpr_mod}{simulated model results (output of simpr::fit)}
}
\value{
tidied model components of fitted simulated data
}
\description{
Turn fitted model of simulated data (output of simpr::fit) into a tidy tibble of model components.
}
\examples{
calc_tidy(simpr_mod)

}
