% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpr_fxns.R
\name{gen}
\alias{gen}
\title{Generate simulated data from specified simpr model (from simpr::variables \%>\% simpr::meta())}
\usage{
gen(simpr, reps)
}
\arguments{
\item{simpr}{simpr class object with metaparamters (e.g., output of simpr::variables() \%>\% simpr::meta())}

\item{reps}{number of replications to run (integer)}
}
\value{
Simulated data
}
\description{
Takes a simpr class object with metaparameters (e.g., output of simpr::variables() \%>\% simpr::meta()) and generates simulated data to be fit with a statistical model (e.g., lm().
}
\examples{
gen(simpr, reps = 1000)

}
