% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include.R
\name{include}
\alias{include}
\title{Include additional postprocessing steps for
simulation}
\usage{
include(obj)
}
\arguments{
\item{obj}{a \code{simpr_meta} object generated
by \code{\link{meta}} or a
\code{simpr_blueprint} object generated by
\code{\link{meta}}}
}
\value{
a \code{simpr_include} object.  This
  can be treated like a \code{simpr_produce}
  object for the purposes of the supported
  functions (see Details).
}
\description{
Allow additional postprocessing steps to be
included in produce_sims(), rather than running
after produce_sims()
}
\details{
Often it is more computationally efficient,
especially when parallel processing, to save
only key results from each simulation rather
than all the data, fit objects, etc.  This
function, run immediately after
\code{\link{blueprint}} or \code{\link{meta}},
allows additional postprocessing steps to be
specified afterwards and thereby included in
the simulation as it proceeds incrementally.

Arbitrary R operations are not supported for
these postprocessing steps; only the
postprocessing steps included in the simpr
package (\code{\link{fit}},
\code{\link{tidy_fits}},
\code{\link{glance_fits}}).
}
