% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpr_fxns.R
\name{meta}
\alias{meta}
\title{Specify metaparameters to vary in simulation}
\usage{
meta(x, ...)
}
\arguments{
\item{x}{simpr class object (e.g., output of variables())}
}
\value{
simpr class object
}
\description{
Takes a simpr class object (e.g., from variables()) and defines the metaparameters for simulation.
}
\examples{
variables(x1 = ~ 2 + rnorm(n),
    x2 = ~ 3 + 2*x1 + rnorm(n, 0, sd = 0.5),
    y = ~ 5 + b1*x1 + b2*x2 + g1*x1*x2 + rnorm(n, 0, sd = s)) \\\%>\\\%
        meta(n = seq(100, 300, by = 20),
             b1 = 1,
             b2 = 1,
             g1 = seq(-1, 1, by = 0.5),
             s = seq(0.2, 50, length.out = 6))

}
