% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpr_spec.R
\name{print.simpr_spec}
\alias{print.simpr_spec}
\title{Methods for simpr_spec class}
\usage{
\method{print}{simpr_spec}(x, ...)
}
\arguments{
\item{x}{a \code{simpr_spec} object}

\item{\dots}{ignored}
}
\description{
Accessor & display methods for simpr_spec class
}
\details{
Class \code{simpr_spec} is created by
\code{\link{blueprint}} and/or
\code{\link{meta}} to specify the simulation
variables, which is produced by
\code{\link{produce}}.
}
