% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpr_fxns.R
\name{variables}
\alias{variables}
\title{Specify functions to generate variables}
\usage{
variables(...)
}
\value{
simpr class object
}
\description{
Creates a simpr class object that specifies formulae for computing each variable in the model.
}
\examples{
variables(x1 = ~ 2 + rnorm(n),
    x2 = ~ 3 + 2*x1 + rnorm(n, 0, sd = 0.5),
    y = ~ 5 + b1*x1 + b2*x2 + g1*x1*x2 + rnorm(n, 0, sd = s))

}
