% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whole_tibble.R
\name{whole_tibble}
\alias{whole_tibble}
\alias{simpr_tibble}
\title{Convert a simpr_sims object back to a
simpr_tibble}
\usage{
whole_tibble(x)
}
\arguments{
\item{x}{A \code{simpr_sims} or \code{simpr_spec} object.}
}
\value{
A tibble with the metaparameters and
  simulation results
}
\description{
Undoes \code{\link{per_sim}} to allow access to simulation
results as a tibble, with simulations available
as a list-column.
}
\details{
This function is the inverse of
\code{\link{per_sim}}. This enables tidyverse
verbs to return to the default behavior of
acting on the full table, as opposed to the
behavior, activated by \code{\link{per_sim}},
of acting elementwise on the simulation
results.
}
