% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_fits.R
\name{apply_fits}
\alias{apply_fits}
\title{Run a given function on a simpr_mod object and
tidy the output.}
\usage{
apply_fits(obj, .f, ..., .progress = FALSE, .options = furrr_options())
}
\arguments{
\item{obj}{tibble with repetition number,
metaparameters, simulated data, and fitted
models, from \code{\link{fit.simpr_tibble}}}

\item{.f}{A function or \code{purrr}-style formula function used for computing on the
fit object}

\item{\dots}{Additional arguments to \code{.f}.}

\item{.progress}{A logical, for whether or not
to print a progress bar for multiprocess,
multisession, and multicore plans .}

\item{.options}{The \code{future} specific
options to use with the workers when using
futures. This must be the result from a call
to
\code{\link[furrr:furrr_options]{furrr_options()}}.}
}
\description{
This function applies a given function to all
fit objects and returns the result in a tidy
tibble.  Any function or \code{purrr}-style
lambda function can be used.
}
\seealso{
\code{\link{tidy_fits}}, \code{\link{glance_fits}}
}
