% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpr_fxns.R
\name{fit}
\alias{fit}
\title{Fit the simulated data (from simpr::gen)}
\usage{
fit(simpr_gen, ...)
}
\arguments{
\item{simpr_gen}{simulated data (output of simpr::gen)}
}
\value{
fitted model object of simulated data
}
\description{
Takes simulated data (output of simpr::gen) and fits to the specified statistical model.
}
\examples{
fit(simpr_gen, lm = ~lm(y ~ x1*x2, data = simpr_gen))

}
