% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen}
\alias{gen}
\title{Generate simulated data from specification}
\usage{
gen(x, reps)
}
\arguments{
\item{x}{a \code{simpr_spec} object generated by \code{\link{variables}} or
\code{\link{meta}}, containing the specifications of the simulation}

\item{reps}{number of replications to run (a whole number greater than 0)}
}
\value{
a \code{simpr_gen} object, which is a tibble with a row for each
  repetition (a total of \code{rep} repetitions) for each combination of
  metaparameters and some extra metadata used by \code{\link{fit}}.  The
  columns are \code{rep} for the repetition number, the names of the
  metaparameters, and a list-column \code{sim_cell} containing the dataset for
  each repetition and metaparameter combination.
}
\description{
Use specification from \code{\link{variables}} or \code{\link{meta}} to
generate simulated data.
}
\details{
This is the third step in the simulation process: after specifying the
variables and metaparameters, \code{gen} is the workhorse function that
actually generates the simulated datasets, one for each replication, for each
combination of metaparameters. You likely want to use the output of
\code{gen} to fit model(s) with \code{\link{fit}}.

Errors you get using this function usually have to do with how you specified
the simulation in  \code{\link{variables}} and \code{\link{meta}}.
}
\examples{
meta_list_out = variables(x = ~ MASS::mvrnorm(n, rep(0, 2), Sigma = S)) \%>\%
  meta(n = c(10, 20, 30),
       S = list(independent = diag(2), correlated = diag(2) + 2)) \%>\%
  gen(3)

 ## View overall structure of the result
 meta_list_out

 ## View an individual dataset of the resulting simulation
 meta_list_out$sim_cell[[1]]

 ## Changing reps will change the number of replications and thus the number of
 ## rows in the output
 meta_list_2 = variables(x = ~ MASS::mvrnorm(n, rep(0, 2), Sigma = S)) \%>\%
  meta(n = c(10, 20, 30),
       S = list(independent = diag(2), correlated = diag(2) + 2)) \%>\%
  gen(4)

 meta_list_2

}
\seealso{
\code{\link{variables}} and \code{\link{meta}} for examples of how
  these functions affect the output of \code{gen}
}
